+++ src/operators/pipeline/sort.ts
@@ -1,6 +1,6 @@
-import { CollationSpec, Options, PipelineOperator } from "../../core/_internal";
+import { Options, PipelineOperator } from "../../core/_internal";
 import { Iterator } from "../../lazy";
-import { Any, AnyObject, Comparator } from "../../types";
+import { Any, AnyObject, CollationSpec, Comparator } from "../../types";
 import {
   assert,
   compare,